+++ tests/modeling/test_modeling.py
@@ -21,6 +21,7 @@ from pharmpy.modeling import (
     has_first_order_elimination,
     has_michaelis_menten_elimination,
     has_mixed_mm_fo_elimination,
+    has_odes,
     has_zero_order_absorption,
     has_zero_order_elimination,
     remove_iiv,
@@ -3527,3 +3528,11 @@ def test_find_volume_parameters_github_issues_1053_and_1062_bis(load_example_mod
     add_peripheral_compartment(model)
     set_michaelis_menten_elimination(model)
     assert find_volume_parameters(model) == _symbols(['V', 'VP1', 'VP2'])
+
+
+def test_has_odes(load_example_model_for_test, datadir, load_model_for_test):
+    model = load_example_model_for_test('pheno')
+    assert has_odes(model)
+    path = datadir / 'minimal.mod'
+    model = load_model_for_test(path)
+    assert not has_odes(model)