+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+next version
+------------
+
+New features
+============
+
+* Add function ``modeling.has_odes``
+
+
 0.89.0 (2023-01-26)
 -------------------